pyproject.toml
[tool.poetry]
name = "drfadvanced"
version = "0.1.0"
description = "Django rest framework advanced concepts with poetry, virtualenv and swagger, followed from https://www.youtube.com/watch?v=83KYtHsUvLY&ab_channel=FaradayAcademy, instruction are available on https://github.com/faraday-academy/django-setup-wiki and https://github.com/mithunadhikari40/django-setup-wiki"
authors = ["Mithun Adhikari"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9"
psycopg2 = "^2.9.2"
djangorestframework = "^3.12.4"
drf-nested-routers = "^0.93.4"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
